refactor: extract server route modules
