+++ apps/server/src/routes/health.ts
@@ -1,0 +1,10 @@
+import type { FastifyInstance } from "fastify";
+
+export function registerHealthRoutes(app: FastifyInstance) {
+  app.get("/health", async () => {
+    return {
+      status: "ok",
+      service: "lowtime-server",
+    };
+  });
+}
